Move rdm/tagged definitions and structures into separated header Signed-off-by:Evgeny Leksikov <evgeny.leksikov@intel.com>
