+++ prov/verbs/src/ep_rdm/verbs_rdm_cm.c
@@ -30,13 +30,13 @@
  * SOFTWARE.
  */
 
-#include <prov/verbs/src/fi_verbs.h>
-
 #include <malloc.h>
 #include <rdma/rdma_cma.h>
 #include <fi_list.h>
 
+#include <prov/verbs/src/fi_verbs.h>
 #include <prov/verbs/src/verbs_utils.h>
+#include <prov/verbs/src/ep_rdm/verbs_rdm.h>
 
 extern struct fi_provider fi_ibv_prov;
 extern struct fi_ibv_rdm_tagged_conn *fi_ibv_rdm_tagged_conn_hash ;
